Add Create Team Type
