+++ src/frontend/src/hooks/design-reviews.hooks.ts
@@ -11,7 +11,8 @@ import {
   getAllDesignReviews,
   getAllTeamTypes,
   getSingleDesignReview,
-  markUserConfirmed
+  markUserConfirmed,
+  createTeamType
 } from '../apis/design-reviews.api';
 import { useCurrentUser } from './users.hooks';
 
@@ -22,6 +23,11 @@ export interface CreateDesignReviewsPayload {
   optionalMemberIds: string[];
   wbsNum: WbsNumber;
   meetingTimes: number[];
+}
+
+export interface CreateTeamTypePayload {
+  name: string;
+  iconName: string;
 }
 
 export const useCreateDesignReviews = () => {
@@ -97,6 +103,27 @@ export const useAllTeamTypes = () => {
     const { data } = await getAllTeamTypes();
     return data;
   });
+};
+
+/**
+ * Custom react hook to create a team type
+ *
+ * @returns the team type created
+ */
+export const useCreateTeamType = () => {
+  const queryClient = useQueryClient();
+  return useMutation<TeamType, Error, CreateTeamTypePayload>(
+    ['teamTypes', 'create'],
+    async (teamTypePayload) => {
+      const { data } = await createTeamType(teamTypePayload);
+      return data;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(['teamTypes']);
+      }
+    }
+  );
 };
 
 /**
